feat: add web stub for plugin loader
