+++ lib/plugins/plugin_loader_web.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+import '../services/service_registry.dart';
+import 'plugin.dart';
+import 'plugin_manager.dart';
+
+class PluginLoader {
+  const PluginLoader();
+
+  List<Plugin> loadBuiltInPlugins() => const [];
+
+  Future<Map<String, bool>> loadConfig() async => <String, bool>{};
+
+  Future<Plugin?> loadFromFile(dynamic file) async => null;
+
+  Future<void> downloadFromUrl(String url, {String? checksum}) async {}
+
+  Future<void> loadAll(
+    ServiceRegistry registry,
+    PluginManager manager, {
+    void Function(double progress)? onProgress,
+    BuildContext? context,
+  }) async {}
+}
